src/api/requirements.txt
requests==2.20.1
adal==1.2.1
Flask==1.0
Flask-API==1.0
Flask-Bootstrap==3.3.7.1
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
azure==4.0.0
azure-common==1.1.20
azure-keyvault==1.1.0
azure-mgmt==4.0.0
azure-mgmt-managementgroups==0.1.0
azure-mgmt-authorization==0.50.0
azure-mgmt-compute==4.6.2
msrestazure==0.4.34
python-dotenv==0.10.1
keyrings.alt==3.1.1
objectpath==0.6.1
azure-cli-core==2.0.70
knack==0.6.3
azure-mgmt-subscription==0.2.0
